requirements.txt
streamlit
yt-dlp
fpdf
whisper
docx2pdf
pydub
spleeter
ffmpeg-python
